refactor(domain, test): random 패키지 생성에 따라, 랜덤 값에 관련있는 객체 패키지 이동
* 패키지 이동에 따라 관련있는 객체의 import 문 수정하고 반영
