+++ src/main/java/lotto/domain/Lottos.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lotto.domain.random.CreateRandomNumbers;
 
 public class Lottos {
