+++ src/test/java/lotto/domain/random/LottoNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.random;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
